Refactor ecosystem integration scanning boundaries
Separate SRM projection, classification policy, and presence aggregation behind the existing scanner facade while preserving ordered integration currency.
